Corrections for redirections, heredocs, leaks, unlinks, and …

+++ srcs/heredoc_files_management_minishell.c
@@ -6,31 +6,40 @@
 /*   By: hauerbac <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/17 13:49:57 by hauerbac          #+#    #+#             */
-/*   Updated: 2024/05/20 16:53:05 by hauerbac         ###   ########.fr       */
+/*   Updated: 2024/05/28 15:55:06 by hauerbac         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "lexer_minishell.h"
+#include "minishell.h"
 
-static int	write_here_doc_file_in_fork(t_cmd *cmd_d)
+static int	open_and_write_here_doc_file_in_fork(t_data *d, t_token *t,
+			t_list *cur)
 {
-	pid_t				pid;
-	int					wstatus;
+	pid_t	pid;
+	int		wstatus;
+	int		res;
 
 	pid = fork();
 	if (pid == -1)
 		return (display_error("Fork error"), -1);
 	if (pid == 0)
-		return (write_here_doc_file(cmd_d));
-	else
 	{
-		if (waitpid(pid, &wstatus, 0) <= 0)
-			return (display_error("waitpid failed"), -2);
-		write(1, "\n", 1);
-		if (WIFEXITED(wstatus) && WEXITSTATUS(wstatus) < 0)
-			return (WEXITSTATUS(wstatus));
-		return (g_exit_status);
+		t->cmd_d->fd1 = open(t->cmd_d->file1, \
+			O_WRONLY | O_CREAT | O_TRUNC, 0600);
+		if (t->cmd_d->fd1 == -1)
+			return (display_error("T.f.cre.e\n"), free_data(d), -4);
+		res = write_here_doc_file(d, t->cmd_d, cur);
+		if (t->cmd_d->fd1 != -1 && close(t->cmd_d->fd1) == -1)
+			return (display_error("T.f.clo.e\n"), free_data(d), -6);
+		t->cmd_d->fd1 = -1;
+		return (empty_dll_before_cur(d->lst, cur, del_el_content), res);
 	}
+	empty_dll_before_cur(d->lst, cur, del_el_content);
+	if (waitpid(pid, &wstatus, 0) <= 0)
+		return (display_error("waitpid failed"), -2);
+	if (g_exit_status == 0 && WIFEXITED(wstatus))
+		return (WEXITSTATUS(wstatus));
+	return (g_exit_status);
 }
 
 static char	*get_file_name_for_heredoc(char *limiter, t_dll_el *prev)
@@ -61,33 +70,30 @@ static char	*get_file_name_for_heredoc(char *limiter, t_dll_el *prev)
 	return (file_name);
 }
 
-int	create_heredoc_file(t_token *t, t_dll_el *prev)
+int	create_heredoc_file(t_data *d, t_token *t, t_dll_el *current)
 {
+	t_list	*curr;
+
 	t->cmd_d->limiter = (char *) malloc((t->src_len + 1) * sizeof(char));
 	if (!t->cmd_d->limiter)
 		return (-3);
 	ft_strlcpy(t->cmd_d->limiter, t->src, t->src_len + 1);
 	t->cmd_d->is_here_doc = 1;
-	t->cmd_d->file1 = get_file_name_for_heredoc(t->cmd_d->limiter, prev);
+	t->cmd_d->file1 = get_file_name_for_heredoc(t->cmd_d->limiter, \
+							current->prev);
 	if (!t->cmd_d->file1)
 		return (free_cmd_d(t->cmd_d), -3);
-	t->cmd_d->fd1 = open(t->cmd_d->file1, O_WRONLY | O_CREAT | O_TRUNC,
-			0644);
-	if (t->cmd_d->fd1 == -1)
-		return (display_err_with_prefix(t->cmd_d->file1, \
-				" temporary file creation error\n"), -4);
-	if (write_here_doc_file_in_fork(t->cmd_d) < 0)
+	curr = (t_list *) current;
+	if (open_and_write_here_doc_file_in_fork(d, t, curr) < 0)
 	{
 		close(t->cmd_d->fd1);
 		t->cmd_d->fd1 = -1;
-		return (display_err_with_prefix(t->cmd_d->file1, \
-				" temporary file writing error\n"), -5);
+		return (display_error("Temporary file writing error\n"), -5);
 	}
-	/*if (close(t->cmd_d->fd1) == -1)
-		return (display_err_with_prefix(t->cmd_d->file1, \
-				" temporary file close error\n"), -6);
-	t->cmd_d->fd1 = -1;*/
-	return (0);
+	if (t->cmd_d->fd1 != -1 && close(t->cmd_d->fd1) == -1)
+		return (perror("Temporary input just created file close error"),
+			-6);
+	return (g_exit_status);
 }
 
 void	close_in_file_and_free_file_name(t_cmd *cmd_d)
@@ -96,9 +102,7 @@ void	close_in_file_and_free_file_name(t_cmd *cmd_d)
 	{
 		if (cmd_d->is_here_doc && cmd_d->fd1 != -1
 			&& close(cmd_d->fd1) == -1)
-		{
 			perror("Temporary input file close error");
-		}
 		else if (!cmd_d->is_here_doc && cmd_d->fd1 != -1
 			&& close(cmd_d->fd1) == -1)
 		{
@@ -111,7 +115,7 @@ void	close_in_file_and_free_file_name(t_cmd *cmd_d)
 		if (cmd_d->is_here_doc && cmd_d->limiter && cmd_d->file1
 			&& access(cmd_d->file1, F_OK) == 0
 			&& access(cmd_d->file1, R_OK) == 0
-			&& unlink(cmd_d->file1) != 0)
+			&& unlink(cmd_d->file1) != 0 && errno != 2)
 			perror("unlink() error");
 		if (cmd_d->file1)
 			free(cmd_d->file1);
